pyproject.toml
[tool.poetry]
name = "neighpy"
version = "0.0.0"
description = ""
authors = ["auggiemarignier <augustin.marignier@anu.edu.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
scipy = "^1.12.0"
joblib = "^1.3.2"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
matplotlib = "^3.8.2"
shapely = "^2.0.2"
pytest-cov = "^4.1.0"
types-tqdm = "^4.66.0.20240106"


[tool.poetry.group.examples.dependencies]
jupyter = "^1.0.0"
matplotlib = "^3.8.2"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-book-theme = "^1.1.2"
myst-nb = "^1.0.0"

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry-dynamic-versioning.substitution]
files = ["neighpy/_version.py"]

[tool.poetry-dynamic-versioning.files."neighpy/_version.py"]
persistent-substitution = true
initial-content = """
  # These version placeholders will be replaced later during substitution.
  __version__ = "0.0.0"
  __version_tuple__ = (0, 0, 0)
"""

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
